Added new folder for views. Moved Portfolio class to Transfer object folder
